# forest-back-demo/docs/integrations.openapi.yaml
openapi: 3.1.0
info:
  title: 산림재난 통합 연계 API
  version: 1.1.0
  description: |
    투비유니콘 통합 플랫폼과 장비·통신망·AI 제공업체 사이의 공통 계약입니다.
    외부 업체는 결과 수신 API를 호출하며, 투비유니콘은 등록된 endpointEnv를 통해
    외부 기능을 호출합니다. 이 명세는 외부 장비·RF·AI 자체 구현을 의미하지 않습니다.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /assets:
    get:
      summary: 등록 장치 목록 조회
      operationId: listAssets
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
        - name: cursor
          in: query
          schema: {type: string}
      responses:
        "200": {description: 장치 목록}
    post:
      summary: 장치 자산 사전등록
      description: 현장 데이터 수신 전에 관리자 권한으로 장치를 등록한다. 자동발견 데이터만으로 자산을 생성하지 않는다.
      operationId: registerAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AssetRegistration"}
      responses:
        "201": {description: REGISTERED 상태의 장치 생성}
        "400": {description: 장치 식별자 또는 기능 목록 오류}
  /assets/{assetId}/status:
    post:
      summary: 장치 운용 상태 변경
      operationId: changeAssetStatus
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [REGISTERED, ACTIVE, SUSPENDED, LOST, RETIRED]
                reason: {type: string}
      responses:
        "200": {description: 상태 변경 및 감사기록 완료}
        "404": {description: 장치 없음}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AssetId:
      name: assetId
      in: path
      required: true
      schema: {type: string, format: uuid}
  schemas:
    AssetRegistration:
      type: object
      required: [assetCode, assetType, serialNumber, capabilities]
      properties:
        assetId: {type: string, format: uuid}
        assetCode: {type: string, maxLength: 100}
        assetType: {type: string, example: PERSONNEL_TERMINAL}
        assetName: {type: string, maxLength: 200}
        ownerOrgCode: {type: string, maxLength: 100}
        modelName: {type: string, maxLength: 100}
        serialNumber: {type: string, maxLength: 200}
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            pattern: "^[A-Z][A-Z0-9_]{1,49}$"
...